==> scripts/01-Get-Pseudobulk.ipynb <==
#### 04-09-2024 This is adapted from Alexandre Pelletier's SEAAD_pseudobulk.ipynb notebook
# Following Alexandre's SEAAD_pseudobulk script

library(Seurat)
library(data.table)
fp<-function(...)file.path(...) # what A.P. is doing here is defining a function fp
# which is a wrapper for the file.path(...) function



# Note: this section is largely based on https://satijalab.org/seurat/articles/de_vignette

#to match with the rds object name, we create a list containing the original cell type name 
cellTypes <- c('Glutamatergic', 'GABAergic', 'Glutamatergic', 'Astrocyte', 'OPC', 
               'Epithelial', 'NPCs', 'VLMC', 'unknown')

# dir w cellTypes .rds files
inputDir <- './outputs/01-Get-Pseudobulk/Seurat/'
outputDir <- './outputs/01-Get-Pseudobulk/'

# for each cell type
for(cellType in cellTypes) {
    # Seurat file for the current cell type
    seuratPath <- fp(dir, paste0(cellType, '.rds'))
    
    # read Seurat file
    seuratObject <- readRDS(seuratPath)
    
    # aggregate expression per genotype
    pseudo_matrix <- AggregateExpression(seuratObject, assays = 'SoupX_counts', 
                                      group.by = 'genotype', return.seurat = FALSE)
    
    # output to csv
    output <- fp(baseDir, paste0(cellType, '_pseudobulk.csv.gz'))
    fwrite(data.table(pseudo_matrix$SoupX_counts, keep.rownames = 'gene_name'), output)
}
